Added a modifier , constructor  and a address variable owner
Constructor Addition:
A constructor has been added to the contract. The constructor initializes the owner variable with the address of the deployer of the contract (msg.sender). This ensures that the deployer is set as the owner when the contract is deployed.

Modifier Inclusion:
The onlyOwner modifier has been introduced. This modifier restricts certain functions, such as the changeTweetLength function, to only be callable by the owner of the contract. This helps enhance security and control over critical contract functionalities.

Tweet Length Flexibility:
The MAX_TWEET_LENGTH variable has been changed from a constant to a public variable. This modification allows the owner of the contract to dynamically adjust the maximum tweet length using the changeTweetLength function. This provides more flexibility in managing the tweet length constraint.